// 03-前端原型demo/B端PC/src/config/menu-catalog.generated.ts
export const generatedSidebarGroupsByModule: Record<string, SidebarGroup[]> = {
// ...
  "device-warning": [
    {
      "id": "物联网iot管理",
      "label": "物联网IOT管理",
      "items": [
        {
          "label": "监控设备",
          "path": "/物联网IOT与预警/物联网IOT管理/监控设备"
        },
        {
          "label": "门禁设备",
          "path": "/物联网IOT与预警/物联网IOT管理/门禁设备"
        },
        {
          "label": "物联设备",
          "path": "/物联网IOT与预警/物联网IOT管理/物联设备"
        },
        {
          "label": "GPS设备",
          "path": "/物联网IOT与预警/物联网IOT管理/GPS设备"
        },
        {
          "label": "人脸配置",
          "path": "/物联网IOT与预警/物联网IOT管理/人脸配置"
        }
      ]
    },
    {
      "id": "预警信息",
      "label": "预警信息",
      "items": [
        {
          "label": "设备预警信息",
          "path": "/物联网IOT与预警/预警信息/设备预警信息"
        },
        {
          "label": "押品预警信息",
          "path": "/物联网IOT与预警/预警信息/押品预警信息"
        },
        {
          "label": "贷中风控管理",
          "path": "/物联网IOT与预警/预警信息/贷中风控管理"
        },
        {
          "label": "风险公示",
          "path": "/物联网IOT与预警/预警信息/风险公示"
        }
      ]
    },
    {
      "id": "预警配置",
      "label": "预警配置",
      "items": [
        {
          "label": "预警等级",
          "path": "/物联网IOT与预警/预警配置/预警等级"
        },
        {
          "label": "设备预警配置",
          "path": "/物联网IOT与预警/预警配置/设备预警配置"
        },
        {
          "label": "订单预警配置",
          "path": "/物联网IOT与预警/预警配置/订单预警配置"
        }
      ]
    }
  ]
} as const